refactor(game): move physics related methods to separate file
